If pipefail isn't set, then non-zero exit codes from `shasum` are masked.
When `shasum` encounters an error, it will print something to stderr, print nothing to stdout, and `cut` happily does its thing with the empty stdout from `shasum` and returns with a zero exit code. Without setting pipefail, the exit code of the rightmost command is used as the status of the entire pipeline, so the conditional check on this line will essentially always pass.

That's bad because $installer_hash is assigned an empty string when this happens. When the empty string is later used as a pattern for `grep -q`, it again ensures that the conditional always passes... and the function announces the checksum has been verified. Yikes.

Realistically, though, that's not as bad as it sounds. If the installer file was corrupted or modified, `shasum` will still do its job and return the hash of the bad file and all the remaining checks work as expected. The edge case described above would only happen as a result of FS/permission issues, or if the installer file is deleted before download_and_install_miniconda() could verify its checksum.